// Project_Apex/frontend/src/api/models/LongTermROIEvent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LongTermROIEvent = {
    id: string;
    createdAt: string;
    planName: string;
    roiPercent: number;
    amount: number;
    description?: string | null;
    investmentId?: string | null;
};